Hopefully fixed the deadlock!
Originally, the semaphore was released in the job before decrementing the
current job count; this could deadlock, because the last semaphore could be
acquired and the job count checked while the job count was still at the old
value, leading to a deadlock (while loop hangs waiting for a release).
I've also fixed a potential problem with accessing the value of job_count
without locking.